requirements.txt
python-dotenv==1.0.1
openai==1.52.2
azure-ai-documentintelligence==1.0.0b3
azure-core==1.30.2
azure-identity==1.17.1
streamlit==1.39.0
python-docx==1.2.0
reportlab==4.4.5